finalized geoshapes and rename them
